DRAFT [morty] apriltag-0.9.8-1.inc: Append to ROS_BUILD_DEPENDS
This component's CMakeLists.txt conditionally adds a demo file to the
build based on whether "opencv" is installed. To make things
deterministic under [morty], append it to do ROS_BUILD_DEPENDS. This
isn't this needed for [thud] because it has a sysroot per recipe, and
"opencv" will never be found.