Revert "DRAFT webos-crystal.mcf: Drop unused meta-webos-backports-2.{7,8} layers"
This reverts commit 56eab249031e692e2e98413c364dfe5cfcf299c7.

The layers aren't empty in OSE 84 (not completely mandatory with a fix for
guider and libmikmod), but in OSE 91 there is also backported curl and for
OSE 138 it's mandatory to provide native mesa), I think we should return it
to webos builds to have the same (or as close as possible) setup as
normal OSE does.

Restore the layers for all ROS Distributions, not just crystal.

Signed-off-by: Martin Jansa <martin.jansa@lge.com>